requirements.txt
fastapi
uvicorn
python-doctr[torch]
pdf2image
openpyxl
ics